% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bio.R
\name{Bio}
\alias{Bio}
\title{Biomass plots}
\usage{
Bio(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{a label for the stock (e.g. CRA1)}

\item{source.dir}{the directory containing the ADMB output files}

\item{target.dir}{the directory to save the plots to}
}
\description{
Biomass plots
}

