% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bio_posterior.R
\name{Bio_posterior}
\alias{Bio_posterior}
\title{Posterior biomass plots}
\usage{
Bio_posterior(stock, source.dir, target.dir = source.dir,
  MCMCOptions = .MCMCOptions, BioOptions = .BioOptions,
  PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{character string: label for the stock (e.g. CRA1)}

\item{source.dir}{character string: the directory containing the ADMB output files}

\item{target.dir}{character string: the directory to save the plots to}

\item{MCMCOptions}{list: mcmc plot options}

\item{PlotOptions}{list: plot options}

\item{BioOption}{list: biomass plot options}
}
\description{
Plots the vulnerable biomass (for Autumn Winter and Spring Summer) and the total biomass over time. Median and the 90% credibility intervals are shown. For the total biomass, 50% intervals are also given.
}
\author{
Charles Edwards
}

