% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioOptions.R
\docType{data}
\name{BioOptions}
\alias{BioOptions}
\title{\describe{
  \item{\code{VulnB = TRUE}}{do the vulnerable biomass plot}
  \item{\code{TotalB = TRUE}}{do the total biomass plot}
  \item{\code{RecB = TRUE}}{do the recruited biomass plot}
  \item{\code{TotalRecB = TRUE}}{do the x plot}
}}
\format{An object of class \code{list} of length 4.}
\usage{
BioOptions
}
\arguments{
\item{ObsPred}{do the observed versus expected CPUE plots}

\item{Resid}{do the residual CPUE plots}
}
\description{
\dontrun{
Bio("CRA5", source.dir = ".", BioOptions = list(VulnB = TRUE, 
                                                 TotalB = TRUE, 
                                                 RecB = TRUE, 
                                                 TotalRecB = TRUE))
}
}
\details{
Biomass plot options

List of options to be passed to biomass plotting functions. This list is loaded with \pkg{RLPlots} and
is called by the functions \code{\link{Bio}}, \code{\link{Bio_posterior}}, \code{\link{MPD.All}} 
and \code{\link{MCMC.All}}. It contains a list of logical values that can be modified by the user 
and passed to the plotting function. Elements in the list include the following logical values:
}
\keyword{datasets}

