% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catches.R
\name{Catches}
\alias{Catches}
\title{Catch plots}
\usage{
Catches(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{character string: label for the stock (e.g. CRA1)}

\item{source.dir}{character string: the directory containing the ADMB output files}

\item{target.dir}{character string: the directory to save the plots to}
}
\description{
Plots the fit of estimated and observed catches for size-limited (legal) and non-size-limited (illegal) catches
}
\author{
Darcy Webber and Charles Edwards
}

