% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Palette.R
\docType{data}
\name{cbPalette2}
\alias{cbPalette2}
\title{The palette with black}
<<<<<<< HEAD
\format{\preformatted{ chr [1:8] "#000000" "#E69F00" "#56B4E9" "#009E73" "#F0E442" ...
=======
\format{\preformatted{ chr [1:8] "#000000" "#E69F00" "#56B4E9" "#009E73" "#F0E442" "#0072B2" "#D55E00" "#CC79A7"
>>>>>>> 80774bd0b6d3423d40e5461db381d6b3cd8e0419
}}
\usage{
cbPalette2
}
\description{
The palette with black
}
\keyword{datasets}

