% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDF_posterior.R
\name{CDF_posterior}
\alias{CDF_posterior}
\title{Cumulative density function plots of the posteriors of parameters}
\usage{
CDF_posterior(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{character string: a label for the stock (e.g. CRA1)}

\item{source.dir}{character string: the source directory. Can be a single directory or a vector
of directories to plot multiple chains.}

\item{target.dir}{character string: the directory to save the plots to}
}
\description{
Use reader.exe to view .psv file and create a more complete parampost.out
file (THK) Traces
}
\author{
Darcy Webber
}

