% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CPUE.R
\name{CPUE}
\alias{CPUE}
\title{CPUE fit plots}
\usage{
CPUE(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{character string: label for the stock (e.g. CRA1)}

\item{source.dir}{character string: the directory containing the ADMB output files}

\item{target.dir}{character string: the directory to save the plots to}
}
\description{
Plots the MPD fit to CPUE data and residual diagnostics
}
\author{
Darcy Webber and Charles Edwards
}

