% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CPUE_posterior.R
\name{CPUE_posterior}
\alias{CPUE_posterior}
\title{CPUE posterior plots}
\usage{
CPUE_posterior(stock, source.dir, target.dir, CPUEOptions = .CPUEOptions,
  MCMCOptions = .MCMCOptions, PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{character string: label for the stock (e.g. CRA1)}

\item{source.dir}{character string: the directory containing the ADMB output files}

\item{target.dir}{character string: optional directory to save the plots to}

\item{CPUEOptions}{list: CPUE options}

\item{MCMCOptions}{list: MCMC options}

\item{PlotOptions}{list: plot options}
}
\description{
Plots the posterior fit to CPUE data and residual diagnostics
}

