% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CPUEOptions.R
\docType{data}
\name{CPUEOptions}
\alias{CPUEOptions}
\title{CPUEOptions}
\format{\preformatted{List of 2
 $ ObsPred: logi TRUE
 $ Resid  : logi TRUE
}}
\usage{
CPUEOptions
}
\arguments{
\item{ObsPred}{do the observed versus expected CPUE plots}

\item{Resid}{do the residual CPUE plots}
}
\description{
Parameters to be passed on to CPUE plotting functions
}
\keyword{datasets}

