% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR.R
\name{CR}
\alias{CR}
\title{CR plots}
\usage{
CR(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{character string: label for the stock (e.g. CRA1)}

\item{source.dir}{character string: the directory containing the ADMB output files}

\item{target.dir}{character string: the directory to save the plots to}
}
\description{
Plots the MPD fit to historic catch rate data and residual diagnostics
}
\author{
Darcy Webber and Charles Edwards
}

