% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CR_posterior.R
\name{CR_posterior}
\alias{CR_posterior}
\title{Catch rate (CR) posterior plots}
\usage{
CR_posterior(stock, source.dir, target.dir = source.dir,
  MCMCOptions = .MCMCOptions, PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{character string: label for the stock (e.g. CRA1)}

\item{source.dir}{character string: the directory containing the ADMB output files}

\item{target.dir}{character string: the directory to save the plots to}

\item{MCMCOptions}{list: MCMC options}

\item{PlotOptions}{list: plot options}
}
\description{
Plots the posterior fit to historic catch rate data and residual diagnostics
}
\author{
Darcy Webber and Charles Edwards
}

