% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CROptions.R
\docType{data}
\name{CROptions}
\alias{CROptions}
\title{CROptions}
\format{An object of class \code{list} of length 3.}
\usage{
CROptions
}
\arguments{
\item{ObsPred}{do the observed versus expected catch rate plots}

\item{Resid}{do the residual catch rate plots}

\item{IsAnnual}{not sure}
}
\description{
Parameters to be passed on to catch rate plotting functions
}
\keyword{datasets}

