% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crosscorrelation.plot.R
\name{crosscorrelation.plot}
\alias{crosscorrelation.plot}
\title{Cross correlation plot}
\usage{
crosscorrelation.plot(x, col = topo.colors(10), ...)
}
\description{
Cross correlation plot
}
\author{
Darcy N. Webber
}

