% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Density_posterior.R
\name{Density_posterior}
\alias{Density_posterior}
\title{Density plots}
\usage{
Density_posterior(stock, source.dir, target.dir = source.dir[1],
  MCMCOptions = .MCMCOptions, PlotOptions = .PlotOptions)
}
\description{
Use reader.exe to view .psv file and create a more complete parampost.out
file (THK) Traces
}
\author{
Darcy Webber
}

