% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Erate.R
\name{Erate}
\alias{Erate}
\title{Exploitation rate plots}
\usage{
Erate(stock, source.dir, target.dir = source.dir,
  PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{source.dir}{the directory containing the ADMB output files}

\item{target.dir}{the directory to save the plots to}

\item{PlotOptions}{plot options}
}
\description{
Exploitation rate by year, season and SL/NSL
}

