% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LFOptions.R
\docType{data}
\name{LFOptions}
\alias{LFOptions}
\title{LFOptions}
\format{\preformatted{List of 14
 $ ObsPred       : logi TRUE
 $ ObsPredYFixed : logi FALSE
 $ Cum           : logi FALSE
 $ StdRes        : logi FALSE
 $ ResPred       : logi TRUE
 $ BoxSize       : logi TRUE
 $ ResQQ         : logi TRUE
 $ QQ            : logi TRUE
 $ Bubble        : logi FALSE
 $ Reshist       : logi TRUE
 $ SamplesPerPage: num 8
 $ SizeLim       : num [1:3] 54 60 60
 $ line          : num 1
 $ sexcode       : int [1:3] 1 2 3
}}
\usage{
LFOptions
}
\description{
LFOptions
}
\keyword{datasets}

