% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mature_posterior.R
\name{Mature_posterior}
\alias{Mature_posterior}
\title{Proportion maturing posterior}
\usage{
Mature_posterior(stock, source.dir, target.dir = source.dir)
}
\description{
Proportion maturing posterior
}
\author{
D'Arcy Webber, Charles Edwards
}

