% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanSize.R
\name{MeanSize}
\alias{MeanSize}
\title{Mean size (tail width, mm) over time}
\usage{
MeanSize(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{source.dir}{the directory where the outputs for a run are sitting}

\item{target.dir}{the directory where the outputs should be saved}
}
\description{
Mean size (tail width, mm) over time
}

