% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_posterior.R
\name{Merge_posterior}
\alias{Merge_posterior}
\title{Merge posterior files and thin}
\usage{
Merge_posterior(source.dir, target.dir = ".", nThin = 2)
}
\description{
Merge posterior files and thin
}
\author{
Darcy Webber
}

