% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Merge_psv.R
\name{Merge_psv}
\alias{Merge_psv}
\title{Merge posterior files and thin}
\usage{
Merge_psv(source.dir, file.names, target.dir = ".", nThin = 2)
}
\description{
Does not work
}
\author{
Darcy Webber
}

