% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpeview.bypar.r
\name{mpeview.bypar}
\alias{mpeview.bypar}
\title{Plot MPE diagnostic output against a 
single control rule parameter}
\usage{
mpeview.bypar(dat, pars, stock, axis.labels, file.suffix = "", target.dir)
}
\description{
Plot MPE diagnostic output against a 
single control rule parameter
}
\author{
Charles Edwards, D'Arcy Webber
}

