% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mpeview.r
\name{mpeview}
\alias{mpeview}
\title{MPE output viewer}
\usage{
mpeview(dat, pars, stock, axis.labels, point.col = "rule",
  label.rules = ifelse(point.col == "", TRUE, FALSE), file.suffix = "plot1",
  target.dir)
}
\arguments{
\item{dat}{a data.frame read in using read.sims.sum}

\item{pars}{the parameters that can be plotted. The first par becomes the x
axis, the remaining pars different panels (choose up to 4 pars)}

\item{stock}{a stock label}

\item{axis.labels}{named list of the form list(x,y) with x and y character vectors containing the x and y labels}

\item{point.col}{the variable used to colour the points, can be coloured by
"rules" or any of the pars (e.g. "par1", "par2", ..., "par10")}

\item{file.suffix}{a suffix to append to the end of the saved plot name}

\item{target.dir}{the directory to save the plot to}
}
\description{
Options for pars that are commonly plotted include the median of the average
commercial catch (AvComm_med), the median of the average recreational catch
(acRec_med), the AAV_med, the median of the average catch per unit effort
(avI_AW_med).
}
\examples{
\dontrun{
 # read in the data
 dat <- read.sims.sum(stock = c("CRA7", "CRA8"))

 # plot all rules for both CRA7 and CRA8
 mpeview(dat, pars = c('AvComm_med', 'AAV_med', 'AvPCPUE_med'), stock = c("CRA7", "CRA8"), axis.labels = list(x = "Commercial Catch", y = c("AAV", "CPUE")))

 # plot all rules for CRA7 only
 mpeview(dat, pars = c('AvComm_med', 'AAV_med', 'AvPCPUE_med'), stock = "CRA7", axis.labels = list(x = "Commercial Catch", y = c("AAV", "CPUE")))
}
}
\author{
Charles Edwards, D'Arcy Webber
}

