% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Multichain.All.R
\name{Multichain.All}
\alias{Multichain.All}
\title{MCMC multichain plots}
\usage{
Multichain.All(stock, source.dir, target.dir = source.dir)
}
\description{
Function which does everything, includes all the individual functions
}
\author{
Darcy Webber
}

