% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotOptions.R
\docType{data}
\name{PlotOptions}
\alias{PlotOptions}
\title{PlotOptions}
\format{\preformatted{List of 12
 $ UsePeriod     : logi FALSE
 $ ModelEndPeriod: num 120
 $ Captions      : logi TRUE
 $ plottype      : chr "png"
 $ plotsize      : num [1:2] 130 130
 $ resolution    : num 100
 $ pointsize     : num 5
 $ theme         : chr "presentation"
 $ colourPalette : chr [1:8] "#000000" "#E69F00" "#56B4E9" "#009E73" ...
 $ thick         : num 2
 $ thin          : num 1
 $ SexNames      : chr [1:3] "Males" "Immature\\nfemales" "Mature\\nfemales"
}}
\usage{
PlotOptions
}
\description{
A list containing options for plotting
}
\keyword{datasets}

