% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POO.R
\name{POO}
\alias{POO}
\title{POO plots}
\usage{
POO(stock, source.dir, target.dir = source.dir, CROptions, PlotOptions)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{CROptions}{}

\item{PlotOptions}{}

\item{run}{the directory where the outputs for a run are sitting}
}
\description{
POO plots
}

