% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POO_posterior.R
\name{POO_posterior}
\alias{POO_posterior}
\title{POO posterior plot}
\usage{
POO_posterior(stock, source.dir, target.dir = source.dir,
  CROptions = .CROptions, PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{PlotOptions}{}

\item{run}{the directory where the outputs for a run are sitting}
}
\description{
POO posterior plot
}

