% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PRIOptions.R
\docType{data}
\name{PRIOptions}
\alias{PRIOptions}
\title{PRIOptions}
\format{\preformatted{List of 2
 $ ObsPred: logi TRUE
 $ Resid  : logi TRUE
}}
\usage{
PRIOptions
}
\description{
PRIOptions
}
\keyword{datasets}

