% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mpe.pars.R
\name{read.mpe.pars}
\alias{read.mpe.pars}
\title{Read harvest control rule pars in MPE control file}
\usage{
read.mpe.pars(stock, source.dir = "./", control.file = "MPE.ctl")
}
\description{
Read harvest control rule pars in MPE control file
}

