% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rec_posterior.R
\name{Rec_posterior}
\alias{Rec_posterior}
\title{Posterior recruitment plots}
\usage{
Rec_posterior(stock, source.dir, target.dir = source.dir,
  MCMCOptions = .MCMCOptions, PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{a label for the stock (e.g. CRA1)}

\item{source.dir}{the directory containing the ADMB output files}

\item{target.dir}{the directory to save the plots to}

\item{MCMCOptions}{mcmc plot options}

\item{PlotOptions}{plot options}
}
\value{
a plot
}
\description{
Posterior recruitment plots
}

