% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule_projection.R
\name{rule_projection}
\alias{rule_projection}
\title{Control rule projection}
\usage{
rule_projection(stock, control.file, source.dir, target.dir = source.dir,
  cpue.range, bins = 60)
}
\arguments{
\item{stock}{a label for the stock (e.g. \code{"CRA1"})}

\item{control.file}{MPE control file (e.g. \code{"Base.mpe"})}

\item{source.dir}{the directory containing the ADMB output files}

\item{target.dir}{the directory to save the plots to}

\item{cpue.range}{named list matching the stock labels that includes a range of offset CPUE values over which to plot}

\item{bins}{number of bins}
}
\value{
a plot
}
\description{
Produces a two dimensional histogram of CPUE/TACC values visited during MP projections
}

