% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select.R
\name{Select}
\alias{Select}
\title{Selectivity}
\usage{
Select(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{run}{the directory where the outputs for a run are sitting}
}
\description{
Selectivity
}

