% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexRatio.R
\name{sexRatio}
\alias{sexRatio}
\title{Sex ratio fit plots}
\usage{
sexRatio(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{run}{the directory where the outputs for a run are sitting}
}
\description{
Sex ratio fit plots
}

