% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SEXROptions.R
\docType{data}
\name{SEXROptions}
\alias{SEXROptions}
\title{SEXROptions}
\format{\preformatted{List of 2
 $ ObsPred: logi TRUE
 $ Resid  : logi TRUE
}}
\usage{
SEXROptions
}
\description{
SEXROptions
}
\keyword{datasets}

