% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Snail.R
\name{Snail}
\alias{Snail}
\title{Snail trail plots}
\usage{
Snail(stock, source.dir = ".", target.dir = source.dir,
  PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{a label for the stock (e.g. CRA1)}

\item{source.dir}{the directory containing the ADMB output files}

\item{target.dir}{the directory to save the plots to}

\item{PlotOptions}{plot options}
}
\description{
Snail trail plots
}

