% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP_posterior.R
\name{SP_posterior}
\alias{SP_posterior}
\title{Posterior surplus production plot}
\usage{
SP_posterior(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{a label for the stock (e.g. CRA1)}

\item{source.dir}{the directory containing the ADMB output files}

\item{target.dir}{the directory to save the plots to}
}
\description{
Posterior surplus production plot
}

