% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SR_posterior.R
\name{SR_posterior}
\alias{SR_posterior}
\title{Sex ratio fit plots}
\usage{
SR_posterior(stock, source.dir, target.dir = source.dir,
  PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{PlotOptions}{}

\item{run}{the directory where the outputs for a run are sitting}
}
\description{
Sex ratio fit plots
}

