% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Tag.R
\name{Tag}
\alias{Tag}
\title{Plot tag fits}
\usage{
Tag(stock, source.dir, target.dir = source.dir, TagOptions = .TagOptions,
  PlotOptions = .PlotOptions)
}
\description{
Tag plots when there are 2 tag files
}

