% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TagOptions.R
\docType{data}
\name{TagOptions}
\alias{TagOptions}
\title{TagOptions}
\format{\preformatted{List of 8
 $ StdResSize : logi TRUE
 $ StdResRel  : logi TRUE
 $ StdResArea : logi TRUE
 $ StdResCond : logi TRUE
 $ StdResType : logi TRUE
 $ Reshist    : logi TRUE
 $ xlim.male  : num [1:2] 40 90
 $ xlim.female: num [1:2] 40 90
}}
\usage{
TagOptions
}
\description{
TagOptions
}
\keyword{datasets}

