% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TraceHisto.R
\name{TraceHisto}
\alias{TraceHisto}
\title{Trace plots and histograms}
\usage{
TraceHisto(stock, source.dir, target.dir = source.dir)
}
\description{
Use reader.exe to view .psv file and create a more complete parampost.out
file (THK) Traces
}

