% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Palette.R
\docType{data}
\name{cbPalette1}
\alias{cbPalette1}
\title{The palette with grey}
\format{\preformatted{ chr [1:8] "#999999" "#E69F00" "#56B4E9" "#009E73" "#F0E442" ...
}}
\usage{
cbPalette1
}
\description{
The palette with grey
}
\keyword{datasets}

