% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palette.R
\docType{data}
\name{cbPalette1}
\alias{cbPalette1}
\title{The palette with grey}
\format{An object of class \code{character} of length 8.}
\usage{
cbPalette1
}
\description{
The palette with grey
}
\keyword{datasets}

