% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palette.R
\docType{data}
\name{cbPalette2}
\alias{cbPalette2}
\title{The palette with black}
\format{An object of class \code{character} of length 8.}
\usage{
cbPalette2
}
\description{
The palette with black
}
\keyword{datasets}

