% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPUEOptions.R
\docType{data}
\name{CPUEOptions}
\alias{CPUEOptions}
\title{CPUEOptions}
\format{An object of class \code{list} of length 2.}
\usage{
CPUEOptions
}
\arguments{
\item{ObsPred}{do the observed versus expected CPUE plots}

\item{Resid}{do the residual CPUE plots}
}
\description{
Parameters to be passed on to CPUE plotting functions
}
\keyword{datasets}

