% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CROptions.R
\docType{data}
\name{CROptions}
\alias{CROptions}
\title{CROptions}
\format{\preformatted{List of 3
 $ ObsPred : logi TRUE
 $ Resid   : logi TRUE
 $ IsAnnual: logi TRUE
}}
\usage{
CROptions
}
\arguments{
\item{ObsPred}{do the observed versus expected catch rate plots}

\item{Resid}{do the residual catch rate plots}

\item{IsAnnual}{not sure}
}
\description{
Parameters to be passed on to catch rate plotting functions
}
\keyword{datasets}

