% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative.quantile.R
\name{cumulative.quantile}
\alias{cumulative.quantile}
\title{Calculate a cumulative quantile}
\usage{
cumulative.quantile(x, probs = 0.5)
}
\description{
Calculate a cumulative quantile
}
\author{
Darcy Webber
}

