% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF.R
\name{LF}
\alias{LF}
\title{Length-frequency function}
\usage{
LF(stock, source.dir, target.dir = source.dir)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{run}{the directory where the outputs for a run are sitting}
}
\description{
To not have size limit of graph LFOptions$SizeLim=0
}

