% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Mature_posterior.R
\name{Mature_posterior}
\alias{Mature_posterior}
\title{Maturity posterior}
\usage{
Mature_posterior(stock, source.dir, target.dir = source.dir,
  PlotOptions = .PlotOptions)
}
\description{
Maturity posterior
}
\author{
Darcy Webber
}

