% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MCMC.All.R
\name{MCMC.All}
\alias{MCMC.All}
\title{MCMC All}
\usage{
MCMC.All(stock, source.dir = ".", target.dir = paste("../", stock, "plots",
  sep = ""), BioOptions = .BioOptions, CROptions = .CROptions,
  PlotOptions = .PlotOptions, MCMCOptions = .MCMCOptions)
}
\description{
Function which does everything, includes all the individual functions
}
\author{
D'Arcy Webber, Charles Edwards
}

