% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mpeview.r
\name{mpeview}
\alias{mpeview}
\title{MPE output viewer}
\usage{
mpeview(dat, pars, stock, axis.labels, point.col = c("rules"),
  file.suffix = "plot1", target.dir, PlotOptions = .PlotOptions)
}
\arguments{
\item{dat}{a data.frame read in using read.sims.sum}

\item{pars}{the parameters that can be plotted. The first par becomes the x
axis, the remaining pars different panels (choose up to 4 pars)}

\item{stock}{a stock label}

\item{axis.labels}{named list of the form list(x,y) with x and y character vectors containing the x and y labels}

\item{point.col}{the variable used to colour the points, can be coloured by
"rules" or any of the pars (e.g. "par1", "par2", ..., "par10")}

\item{file.suffix}{a suffix to append to the end of the saved plot name}

\item{target.dir}{the directory to save the plot to}

\item{PlotOptions}{plot options}
}
\description{
Options for pars that are commonly plotted include the median of the average
commercial catch (AvComm_med), the median of the average recreational catch
(acRec_med), the AAV_med, the median of the average catch per unit effort
(AvI_AW_med)
}
\author{
Charles Edwards, Darcy Webber
}

