% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOptions.R
\docType{data}
\name{PlotOptions}
\alias{PlotOptions}
\title{PlotOptions}
\format{An object of class \code{list} of length 12.}
\usage{
PlotOptions
}
\description{
A list containing options for plotting
}
\keyword{datasets}

