% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotType.R
\name{PlotType}
\alias{PlotType}
\title{Specify the type of plot you would like to create}
\usage{
PlotType(filename, width, height)
}
\arguments{
\item{filename}{the directory and name of the file to save}

\item{width}{the width of the plot in mm}

\item{height}{the height of the plot in mm}
}
\value{
a plot device
}
\description{
Specify the type of plot you would like to create
}

