% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QQpost.R
\name{QQpost}
\alias{QQpost}
\title{QQpost}
\usage{
QQpost(data, ylab = "Sample quantiles", xlab = "Theoretical quantiles",
  useminresid = FALSE, minresid = 0)
}
\description{
function to make QQ plot from posterior
}

