% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Select.R
\name{Select}
\alias{Select}
\title{Selectivity}
\usage{
Select(stock, source.dir, target.dir = source.dir,
  PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{PlotOptions}{}

\item{run}{the directory where the outputs for a run are sitting}
}
\description{
Selectivity
}

