% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Select_posterior.R
\name{Select_posterior}
\alias{Select_posterior}
\title{Selectivity posterior}
\usage{
Select_posterior(stock, source.dir, target.dir = source.dir,
  PlotOptions = .PlotOptions)
}
\description{
Selectivity posterior
}
\author{
Darcy Webber
}

