% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sexRatio.R
\name{sexRatio}
\alias{sexRatio}
\title{sex ratio fit plots}
\usage{
sexRatio(stock, source.dir, target.dir = source.dir,
  SEXROptions = .SEXROptions, PlotOptions = .PlotOptions)
}
\arguments{
\item{stock}{the name of the stock (e.g. CRA1)}

\item{SEXROptions}{}

\item{PlotOptions}{}

\item{run}{the directory where the outputs for a run are sitting}
}
\description{
sex ratio fit plots
}

