% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Trace_posterior.R
\name{Trace_posterior}
\alias{Trace_posterior}
\title{Trace plots}
\usage{
Trace_posterior(stock, source.dir, target.dir = source.dir[1],
  MCMCOptions = .MCMCOptions, PlotOptions = .PlotOptions)
}
\description{
Use reader.exe to view .psv file and create a more complete parampost.out
file (THK) Traces
}

